Check slamdata dependencies for conflicts on Travis
If building on Travis, the root project will run the "versionCheck"
task.

Version check is performed by adding the sbt-coursier plugin and
adding the slamdata organization for strict checking. The coursier
plugin requires the coursier sbt launcher when using sbt 1.3, so
the coursier sbt runner is downloaded, if not present, and used to
start a new instance of sbt to run the above.

After running version check, "./sbt update" is executed to return
sbt project files to their original state, before finishing the
version check task.

The version check task is configured to run exclusively, to avoid
issues.