Incorporate code review feedback
"in Global" on globalSettings is (probably) redudant.

For start-of-line on regex, though it is being used in a way that
already enforces that. That makes the regex mandate it, instead of
the usage site.

Switch from wget to curl, as it is more widely pre-installed.

Remove method declaration which is followed by its sole use.